feat: deserialize packet

+++ backend/pkg/network/packet.go
@@ -1,5 +1,13 @@
 package network
 
+// PACKET STRUCTURE
+//+---------+------+------+------+
+//| Version | Type | Size | Data |
+//+---------+------+------+------+
+//|       1 |    1 |    2 | 1P   |
+//+---------+------+------+------+
+//
+
 const (
 	PacketVersion1 = iota + 1
 )
@@ -14,6 +22,6 @@ const (
 type TCPPacket struct {
 	Version uint8
 	Type    uint8
-	Size    uint8
+	Size    uint16
 	Data    []byte
 }
